pet_feeder/pet_feeder/requirements.txt
flask==3.0.0
flask-cors==4.0.0
paho-mqtt==1.6.1
sqlalchemy==2.0.23
python-dotenv==1.0.0
psycopg2-binary==2.9.9
pandas==2.1.4
numpy==1.26.2
python-dateutil==2.8.2
requests==2.31.0
